% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesColour.R
\name{aesColour}
\alias{aesColour}
\title{aesColor}
\usage{
aesColour(type)
}
\arguments{
\item{type}{character of label and inputId of element}
}
\value{
UI object
}
\description{
ColorInput UI production for discrete variables.
}
\keyword{internal}
