% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesColourContNS.R
\name{aesColourContNS}
\alias{aesColourContNS}
\title{aesColorContNS}
\usage{
aesColourContNS(type, session)
}
\arguments{
\item{type}{character of label and inputId of element}

\item{session}{shiny session object}
}
\value{
UI object
}
\description{
ColorInput UI production for continuous variables for module.
}
\keyword{internal}
