% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesColourNS.R
\name{aesColourNS}
\alias{aesColourNS}
\title{aesColorNS}
\usage{
aesColourNS(type, session)
}
\arguments{
\item{type}{character of label and inputId of element}

\item{session}{shiny session object}
}
\value{
UI object
}
\description{
ColorInput UI production for discrete variables for module.
}
