% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesSelect.R
\name{aesSelect}
\alias{aesSelect}
\title{aesSelect}
\usage{
aesSelect(type)
}
\arguments{
\item{type}{character of label and inputId of element}
}
\value{
UI object
}
\description{
SelectizeInput UI production for palletes.
}
\keyword{internal}
