% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesSelectNS.R
\name{aesSelectNS}
\alias{aesSelectNS}
\title{aesSelectNS}
\usage{
aesSelectNS(type, session)
}
\arguments{
\item{type}{character of label and inputId of element}

\item{session}{shiny session object}
}
\value{
UI object
}
\description{
SelectizeInput UI production for palletes for module.
}
\keyword{internal}
