% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesSlide.R
\name{aesSlide}
\alias{aesSlide}
\title{aesSlide}
\usage{
aesSlide(type)
}
\arguments{
\item{type}{character of label and inputId of element}
}
\value{
UI object
}
\description{
sliderInput UI production for continuous variables
}
\keyword{internal}
