% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesSlideNS.R
\name{aesSlideNS}
\alias{aesSlideNS}
\title{aesSlideNS}
\usage{
aesSlideNS(type, session)
}
\arguments{
\item{type}{character of label and inputId of element}

\item{session}{shiny session object}
}
\value{
UI object
}
\description{
sliderInput UI production for continuous variables for module
}
\keyword{internal}
