% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argValue.R
\name{arg.value}
\alias{arg.value}
\title{arg.value}
\usage{
arg.value(item, obj.elems)
}
\arguments{
\item{item}{character type of gg layer}

\item{obj.elems}{elements in layer}
}
\value{
list
}
\description{
Creates list of arguments and values in gg layer.
}
\keyword{internal}
