% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argValueNS.R
\name{arg.valueNS}
\alias{arg.valueNS}
\title{arg.valueNS}
\usage{
arg.valueNS(item, obj.elems, session)
}
\arguments{
\item{item}{character type of gg layer}

\item{obj.elems}{elements in layer}

\item{session}{shiny session object}
}
\value{
list
}
\description{
Creates list of arguments and values in gg layer for module.
}
\keyword{internal}
