% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ggedit.R
\name{as.ggedit}
\alias{as.ggedit}
\title{Try to coerce a ggplot object into a ggedit object}
\usage{
as.ggedit(p)
}
\arguments{
\item{p}{an object}
}
\value{
an object of class ggedit
}
\description{
Applied to ggplot objects to use the ploting function of ggedit
}
\examples{
p=ggplot(iris,aes(x =Sepal.Length,y=Sepal.Width))
p1=p+geom_point(aes(colour=Species))+geom_line()
p2=p+geom_point()+geom_smooth(method='loess')
p3=list(p1,p2)
p4=as.ggedit(p3)
plot(p4)
}
\seealso{
\code{\link{plot.ggedit}},\code{\link[ggplot2]{ggplot}}
}

