% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.gglist.R
\name{as.gglist}
\alias{as.gglist}
\alias{as.gglist.gglist}
\alias{as.gglist.list}
\alias{as.gglist.ggplot}
\alias{as.gglist.ggmatrix}
\title{recasts to gglist}
\usage{
as.gglist(plot)

\method{as.gglist}{gglist}(plot)

\method{as.gglist}{list}(plot)

\method{as.gglist}{ggplot}(plot)

\method{as.gglist}{ggmatrix}(plot)
}
\arguments{
\item{plot}{object}
}
\value{
gglist object
}
\description{
adds gglist class to object
}
