% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_layer.R
\name{class_layer}
\alias{class_layer}
\title{class_layer}
\usage{
class_layer(p)
}
\arguments{
\item{p}{ggplot}
}
\value{
data.frame
}
\description{
Determines the class of the arguments in a gg layer.
}
\keyword{internal}
