% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneLayer.R
\name{cloneLayer}
\alias{cloneLayer}
\title{Creates an independent copy of a ggplot object layer}
\usage{
cloneLayer(l)
}
\arguments{
\item{l}{ggplot2 object layer}
}
\value{
ggproto object
}
\description{
Creates copies of ggplot layers from within ggplot objects
 that are independent of the parent object.
}
\details{
ggplot objects are comprimsed of layer objects. Once compiled they 
are part of the plot object environment and if they are changed internally regardless of 
where they are in the (ie different environment) it will change the original plot. This function allows to 
create replicates of the plot layers and edit them independent of the original plot.
}
\examples{
p=ggplot(iris,aes(x =Sepal.Length,y=Sepal.Width))
p=p+geom_point(aes(colour=Species))+geom_line()
p$layers[[1]]
newLayer=cloneLayer(l=p$layers[[1]])
all.equal(p$layers[[1]],newLayer)
}

