% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_root.R
\name{cloneRoot}
\alias{cloneRoot}
\title{clone root of ggplot object}
\usage{
cloneRoot(obj, verbose = FALSE)
}
\arguments{
\item{obj}{ggplot object}

\item{verbose}{boolean, toggles to return object or script (TRUE), Default: FALSE}
}
\value{
gg object or script
}
\description{
clone root of ggplot object and return either a new gg object or
a string to parse and evaluate
}
\examples{
 cloneRoot(pList$pointSmooth)
 cloneRoot(pList$pointSmooth,verbose=TRUE)
}
\seealso{
\code{\link[utils]{capture.output}}
}
