% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloneScales.R
\name{cloneScales}
\alias{cloneScales}
\title{Clone ggplot2 scales from compiled ggplot objects}
\usage{
cloneScales(p, verbose = FALSE)
}
\arguments{
\item{p}{ggplot object}

\item{verbose}{boolean, if TRUE then returns script to eval(parse) if FALSE returns new compiled object Default: FALSE}
}
\value{
ggplot scale or script (depends on verbose)
}
\description{
Clone ggplot2 scales from compiled ggplot objects returns not 
the original call but the nested call
}
\examples{
 # p <- pList$pointSmooth+scale_colour_continuous(low='red')
 # p
 # pList$pointSmooth+cloneScales(p)
}
