% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(e1, e2, verbose = TRUE)
}
\arguments{
\item{e1}{theme object}

\item{e2}{theme object}

\item{verbose}{logical to control if the output is a character of script or a theme object (default TRUE)}
}
\value{
theme object or character depending on verbose
}
\description{
Compare differences theme object e1 (new theme) to theme object e2 (old theme)
}
\examples{
compare(theme_bw(),theme_get())
compare(theme_bw(),theme_get(),verbose=FALSE)
}
