% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_aes_ggplotBuild.R
\name{fetch_aes_ggplotBuild}
\alias{fetch_aes_ggplotBuild}
\title{fetch_aes_ggplotBuild}
\usage{
fetch_aes_ggplotBuild(p, geom_list)
}
\arguments{
\item{p}{ggplot}

\item{geom_list}{character vector of geoms in plot}
}
\description{
Converts ggplot layers to list.
}
\keyword{internal}
