% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_aes_layer.R
\name{fetch_layer_aes}
\alias{fetch_layer_aes}
\title{fetch_layer_aes}
\usage{
fetch_layer_aes(self, data = data.frame(x = 1), params = list(),
  stats = list())
}
\description{
Checks the aesthetics in a layer for internal use only.
}
\keyword{internal}
