% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontDefaults-data.R
\docType{data}
\name{fontDefaults}
\alias{fontDefaults}
\title{fontDefaults}
\format{An object of class \code{"character"}}
\usage{
fontDefaults
}
\description{
Character of fonts used as options in ggedit.
}
\examples{
data(fontDefaults)
fontDefaults
}
\keyword{datasets}
