% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_list.R
\name{geom_list}
\alias{geom_list}
\title{geom_list}
\usage{
geom_list(p)
}
\arguments{
\item{p}{ggplot}
}
\value{
character vector
}
\description{
Returns character vector of geom types in ggplot.
}
\keyword{internal}
