% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_session.R
\name{gg_session}
\alias{gg_session}
\title{retreive all functions that create ggroto layers or stats}
\usage{
gg_session(gg_pkg = NULL)
}
\arguments{
\item{gg_pkg}{character, package names to search in if NULL an internal 
search will look in loaded packages, Default: NULL}
}
\value{
data.frame
}
\description{
Retreive all functions that create ggroto layers or stats
}
\examples{
gg_session('ggplot2')

}
