% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_vetting.R
\name{gg_vetting}
\alias{gg_vetting}
\title{Backcheck what functions created the layers in a ggplot2 plot object}
\usage{
gg_vetting(p, obj = ggedit_opts$get("session_geoms"))
}
\arguments{
\item{p}{gg, compiled ggplot object}

\item{obj}{data.frame, contains the mapping of layer functions as created in gg_session(),
Default: ggedit_opts$get('session_geoms')}
}
\value{
data.frame
}
\description{
Validate geoms with their unique attributes
}
\examples{
gg_vetting(pList$boxplotWrap)
lapply(pList,gg_vetting)
}
