% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggedit.opts-data.R
\docType{data}
\name{ggedit.opts}
\alias{ggedit.opts}
\title{list that containts options for ggedit}
\format{A data frame with 73 rows and 5 variables:
\describe{
  \item{\code{fn}}{character function call}
  \item{\code{geom}}{character underlying Geom}
  \item{\code{position}}{character underlying Position}
  \item{\code{stat}}{character underlying Stat}
  \item{\code{pkg}}{character the namespace the function belongs to} 
}}
\usage{
ggedit.opts
}
\description{
list currently includes a dataframe geom_opts that maps protos to geom calls
}
\keyword{datasets}
