% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggedit.R
\name{ggedit}
\alias{ggedit}
\title{Interactive shiny widget for editing ggplot layers and themes}
\usage{
ggedit(p.in, viewer = paneViewer(minHeight = 1000), ...)
}
\arguments{
\item{p.in}{ggplot2 plot object or list of objects}

\item{viewer}{shiny viewer options}

\item{idx}{numeric of which index of geom to remove}
}
\description{
Control where the gadget is displayed in RStudio (pane,dialog,browser)
}
\examples{
p=ggplot(iris,aes(x =Sepal.Length,y=Sepal.Width))
p=p+geom_point(aes(colour=Species))+geom_line()
pnew=ggedit(p)
pnew
}

