% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeditModule.R
\name{ggEdit_Module}
\alias{ggEdit_Module}
\alias{ggEdit}
\title{ggEdit}
\usage{
ggEdit(input, output, session, obj, verbose = T, showDefaults = F)
}
\arguments{
\item{input}{shinyapp input argument}

\item{output}{shinyapp output argument}

\item{session}{shinyapp session argument}

\item{obj}{ggplot as reactive shiny object}

\item{verbose}{logical to control if the output includes script for layers and themes calls for parsing to create objects (default, verbose=F)}

\item{showDefaults}{toggle to control if the verbose output shows all the input arguments passed to the proto object (if verbose==FALSE then ignored)}
}
\description{
The ggedit module.
}
\keyword{internal}
