% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.ggedit.R
\name{is.ggedit}
\alias{is.ggedit}
\title{Is the object of class ggedit}
\usage{
is.ggedit(p)
}
\arguments{
\item{p}{an object}
}
\value{
logical - is the object of class ggedit
}
\description{
Is the object of class ggedit. Very basic for many functions
 in the package.
}
