% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.NullOb.R
\name{is.NullOb}
\alias{is.NullOb}
\title{is.NullOb}
\usage{
is.NullOb(x)
}
\arguments{
\item{x}{vector or data.frame}
}
\value{
logical
}
\description{
Checks if any columns have a null in them.
}
\keyword{internal}
