% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layersList.R
\name{layersList}
\alias{layersList}
\title{layersList}
\usage{
layersList(obj)
}
\arguments{
\item{obj}{ggplot2 plot object or list of plot objects}
}
\value{
list of aesthetics and their values for each layer in a plot
}
\description{
Runs the ggplot_build function on the input and converts
 the output data objects into a nested list with the unique values of 
 each of the aesthetic columns.
}
\examples{
p=ggplot2::ggplot(iris,ggplot2::aes(x=Sepal.Length,y=Sepal.Width))
p=p+ggplot2::geom_point(ggplot2::aes(colour=Species))+ggplot2::geom_line()
p
p.list=layersList(p)
p.list

}
