% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{pList}
\alias{pList}
\title{List of plots for ggedit examples}
\format{An object of class \code{"list"}}
\usage{
pList
}
\description{
List of ggplot objects for examples in ggedit.
}
\details{
list includes a geom_point, geom_point+facet_wrap, geom_boxplot+facet_wrap,
geom_point+geom_line, geom_point+geom_smooth, geom_point+geom_line+facet_wrap,
geom_point+geom_line+facet_grid
}
\keyword{datasets}
