% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ggedit.R
\name{plot.ggedit}
\alias{plot.ggedit}
\title{Plots ggedit plot output objects}
\usage{
\method{plot}{ggedit}(obj, plot.layout = NULL)
}
\arguments{
\item{obj}{ggedit output of list of plots from the ggedit output}

\item{plot.layout}{list which defines the location of each plot within
the viewport layout. (see details)}
}
\details{
If plot.layout is NULL then a default layout is used filling in
a grid.layout row.wise. Defining plot.layout as a nested list giving 
the rows and the column indicies of each plot will produce a specific layout.
}
\examples{
p=ggplot(iris,aes(x =Sepal.Length,y=Sepal.Width))
p=p+geom_point(aes(colour=Species))+geom_line()
p
x=ggedit(p)
plot(x)
plot(x$UpdatedPlots)
p1=pList[c(1:2)]
x1=ggedit(p1)
plot(x1)
plot(x1,plot.layout = list(list(rows=2,cols=2),list(rows=1,cols=1:2)))
}

