% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto_features.R
\name{proto_features}
\alias{proto_features}
\title{ggplot2 layer proto extraction}
\usage{
proto_features(l)
}
\arguments{
\item{l}{ggproto}
}
\value{
data.frame
}
\description{
Extract geom, stat and position protos from a ggplot2 layer
}
\examples{
proto_features(ggplot2::geom_smooth())
proto_features(ggplot2::annotation_logticks())

}
