% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_geom.R
\name{remove_geom}
\alias{remove_geom}
\title{Remove a layer from a compiled ggplot2 object.}
\usage{
remove_geom(p, geom, idx)
}
\arguments{
\item{p}{ggplot2 plot object}

\item{geom}{character string of the name of the layer to remove}

\item{idx}{numeric of which index of geom to remove}
}
\description{
Removes specified layers from a ggplot object.
}
\examples{
p <- ggplot2::ggplot(iris,ggplot2::aes(x =Sepal.Length,y=Sepal.Width))
p <- p+ggplot2::geom_point(ggplot2::aes(colour=Species))+ggplot2::geom_line()
p
pnew <- p\%>\%remove_geom('point',1)
pnew
}
