% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideDefaults-data.R
\docType{data}
\name{slideDefaults}
\alias{slideDefaults}
\title{slideDefaults}
\format{An object of class \code{"list"}}
\usage{
slideDefaults
}
\description{
List of limits for layer aesthetics.
}
\examples{
data(slideDefaults)
slideDefaults
}
\keyword{datasets}
