% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ggedit.R
\name{summary.ggedit}
\alias{summary.ggedit}
\title{Print to console verbose outputs of objects of class ggedit}
\usage{
\method{summary}{ggedit}(object, ...)
}
\arguments{
\item{object}{ggedit object}

\item{...}{used to pass show.structure}
}
\description{
function to tidy the ggedit output to single script calls for each plot
}
\details{
use show.structure (boolean) to control if the output shows the full data.frame
structure or just a placeholder [data.frame], Default: FALSE
}
