% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeDefaultClass-data.R
\docType{data}
\name{ThemeDefaultClass}
\alias{ThemeDefaultClass}
\title{ThemeDefaultClass}
\format{An object of class \code{"data.frame"}}
\usage{
ThemeDefaultClass
}
\description{
Data.frame of default classes for theme elements.
}
\examples{
data(ThemeDefaultClass)
ThemeDefaultClass
}
\keyword{datasets}
