% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeFetch.R
\name{themeFetch}
\alias{themeFetch}
\title{themeFetch}
\usage{
themeFetch(a = theme_get())
}
\arguments{
\item{a}{gg theme}
}
\value{
list
}
\description{
returns a nested list of a gg theme object, for ggedit gadget or module.
}
\keyword{internal}
