% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeFetchFull.R
\name{themeFetchFull}
\alias{themeFetchFull}
\title{themeFetchFull}
\usage{
themeFetchFull(a = theme_get())
}
\arguments{
\item{a}{gg theme}
}
\value{
list
}
\description{
Returns a nested list of all elements in gg theme object, 
for plot.theme plot.
}
\keyword{internal}
