% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeListDepth.R
\name{themeListDepth}
\alias{themeListDepth}
\title{themeListDepth}
\usage{
themeListDepth(this, thisdepth = 0)
}
\arguments{
\item{this}{list}

\item{thisdepth}{numeric of current depth in recursive call}
}
\value{
numeric
}
\description{
Returns depth of nested list.
}
\keyword{internal}
