% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeMakePanel.R
\name{themeMakePanel}
\alias{themeMakePanel}
\title{themeMakePanel}
\usage{
themeMakePanel(x, y = NULL, session = NULL)
}
\arguments{
\item{x}{list from themeFetch}

\item{y}{character, names of nested themes elements, Default: NULL}

\item{session}{shiny session object}
}
\value{
UI
}
\description{
BS modal production for the gg theme.
}
\keyword{internal}
