% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeMakePanelNS.R
\name{themeMakePanelNS}
\alias{themeMakePanelNS}
\title{themeMakePanel}
\usage{
themeMakePanelNS(x, y = NULL, session)
}
\arguments{
\item{x}{list from themeFetch}

\item{session}{shiny session object}
}
\value{
UI
}
\description{
BS modal production for the gg theme, for module.
}
\keyword{internal}
