% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePanel.R
\name{themePanel}
\alias{themePanel}
\title{themePanel}
\usage{
themePanel(obj.in, obj.name, session = NULL)
}
\arguments{
\item{obj.in}{theme object in list form}

\item{obj.name}{name of theme object}

\item{session}{shiny session object}
}
\value{
UI
}
\description{
UI production for the gg theme.
}
\keyword{internal}
