% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThemeTips-data.R
\docType{data}
\name{themeTips}
\alias{themeTips}
\title{themeTips}
\format{An object of class \code{"list"}}
\usage{
data(themeTips)
}
\description{
List of HTML tips for the theme BS modal.
}
\examples{
data(themeTips)
themeTips
}
\keyword{datasets}
