% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplayout.R
\name{vplayout}
\alias{vplayout}
\title{vplayout}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{numeric of row number}

\item{y}{numeric of column number}
}
\value{
viewport
}
\description{
Viewport with layout.
}
\keyword{internal}
