% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesColourCont.R
\name{aesColourCont}
\alias{aesColourCont}
\title{aesColorCont}
\usage{
aesColourCont(type)
}
\arguments{
\item{type}{character of label and inputId of element}
}
\value{
UI object
}
\description{
ColorInput UI production for continuous variables.
}
\keyword{internal}
