% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ggedit.R
\name{as.ggedit}
\alias{as.ggedit}
\title{Try to coerce a ggplot object into a ggedit object}
\usage{
as.ggedit(plot)
}
\arguments{
\item{plot}{an object}
}
\value{
an object of class ggedit
}
\description{
Applied to ggplot objects to use the ploting function of ggedit.
}
\examples{

p  <- ggplot2::ggplot(iris,ggplot2::aes(x =Sepal.Length,y=Sepal.Width))

p1 <- p + 
ggplot2::geom_point(ggplot2::aes(colour=Species)) + 
ggplot2::geom_line()

p2 <- p + 
ggplot2::geom_point() + 
ggplot2::geom_smooth(method='loess')

p3 <- list(p1,p2)

p4 <- as.ggedit(p3)

p4
}
\seealso{
\code{\link{print.ggedit}}, \code{\link[ggplot2]{ggplot}}
}
