% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_vetting.R
\name{gg_vetting}
\alias{gg_vetting}
\title{Backcheck what functions created the layers in a ggplot2 plot object}
\usage{
gg_vetting(p, obj)
}
\arguments{
\item{p}{gg, compiled ggplot object}

\item{obj}{data.frame, contains the mapping of layer functions as created in gg_session()}
}
\value{
data.frame
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
x<-gg_session('ggplot2')
p=pList$boxplotWrap
gg_vetting(p,x)
lapply(pList,gg_vetting,obj=x)
}
