% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggeditModule.R
\name{ggEdit_Module}
\alias{ggEdit_Module}
\alias{ggEdit}
\title{ggEdit}
\usage{
ggEdit(input, output, session, obj)
}
\arguments{
\item{input}{shinyapp input argument}

\item{output}{shinyapp output argument}

\item{session}{shinyapp session argument}

\item{obj}{ggplot as reactive shiny object}
}
\description{
The ggedit module.
}
\keyword{internal}
