% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layersList.R
\name{layersList}
\alias{layersList}
\title{Convert aesthetics of layers of a ggplot to a list.}
\usage{
layersList(obj)
}
\arguments{
\item{obj}{ggplot2 plot object or list of plot objects}
}
\value{
list of aesthetics and their values for each layer in a plot
}
\description{
Runs the ggplot_build function on the input and converts the output 
data objects into a nested list with the unique values of each of the aesthetic columns
}
\examples{
p=ggplot(iris,aes(x =Sepal.Length,y=Sepal.Width))
p=p+geom_point(aes(colour=Species))+geom_line()
p
p.list=layersList(p)
p.list

}

