% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.theme.R
\name{plot.theme}
\alias{plot.theme}
\title{Plots theme elements in tile layout}
\usage{
\method{plot}{theme}(obj, fnt = 11)
}
\arguments{
\item{obj}{ggplot theme}

\item{fnt}{numeric font size of text in plot}
}
\examples{
plot(theme_bw(),fnt=10)
}

