% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ggedit.R
\name{print.ggedit}
\alias{print.ggedit}
\title{Print verbose outputs of objects of class ggedit}
\usage{
\method{print}{ggedit}(x, ...)
}
\arguments{
\item{x}{ggedit object}

\item{...}{nothing}
}
\description{
function to tidy the ggedit output to single script calls for each plot
}
