% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ggedit.R
\name{print.ggedit}
\alias{print.ggedit}
\title{Print ggedit objects}
\usage{
\method{print}{ggedit}(x, ...)
}
\arguments{
\item{x}{list of ggplot2 plot objects}

\item{...}{list which defines the location of each plot within
the viewport layout. (see details)}
}
\description{
Plots lists of ggplot2 plot objects
layout functionality.
}
\details{
If ... is NULL then a default layout is used filling in
a grid.layout row.wise. Defining plot.layout as a nested list giving 
the rows and the column indicies of each plot will produce a specific layout.
}
\examples{
\donttest{
p <- as.gglist(list(pList[[1]],pList[[2]]))
p

p1 <- p+geom_hline(aes(yintercept=3))
p1

print(p1,plot.layout = list(list(rows=2,cols=2),list(rows=1,cols=1:2)))
}
}
