% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ggedit.R
\name{summary.ggedit}
\alias{summary.ggedit}
\title{Print to console verbose outputs of objects of class ggedit}
\usage{
\method{summary}{ggedit}(x, show.structure = FALSE, ...)
}
\arguments{
\item{x}{ggedit object}

\item{show.structure}{boolean, controls if the output shows the full data.frame structure or just a placeholder [data.frame], Default: FALSE}

\item{...}{nothing}
}
\description{
function to tidy the ggedit output to single script calls for each plot
}
\details{
To see the complete structure of the data.frame
}
