% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeMakePanel.R
\name{themeMakePanel}
\alias{themeMakePanel}
\title{themeMakePanel}
\usage{
themeMakePanel(x, y = NULL)
}
\arguments{
\item{x}{list from themeFetch}
}
\value{
UI
}
\description{
BS modal production for the gg theme.
}
\keyword{internal}
